requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
boto3==1.24.2
botocore==1.27.2
Django==4.0.4
django-crispy-forms==1.14.0
django-storages==1.12.3
jmespath==1.0.0
Pillow==9.1.1
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9